refactor: component에서 service 로직 분리하기
