+++ docs/REQUIREMENTS-STEP2.md
@@ -43,5 +43,5 @@
 - [x] css 컴포넌트 별로 나누기
 - [x] css 이름 일관성 맞추기
 - [x] js 셀렉팅은 id로, css 셀렉팅은 class 하기
-- [ ] 컴포넌트에서 service 로직 분리하기
+- [x] 컴포넌트에서 service 로직 분리하기
 - [ ] 구조 그려서 설명(옵저버 패턴에 대해 + (구체적으로) state와 component와의 관계도)
